+++ crates/dawn-sector/src/transit.rs
@@ -333,7 +333,7 @@ mod tests {
         let (raft, mut rx) = raft_handle();
         let gate = *node.jump_gate(JumpGateId(0)).expect("Sector 0 has Gate 0");
         let near_gate_abs = [
-            gate.abs_m[0] - (gate.activation_radius as f64 * 0.5),
+            gate.abs_m[0] - (gate.activation_radius * 0.5),
             gate.abs_m[1],
             gate.abs_m[2],
         ];
